feat: add disabled buttons tate
